requirements.txt
Flask
flask-login
flask-SQLAlchemy
requests
python-dotenv
psycopg2-binary